Fixed renderer logger not working well, some logs duplicating, some inconsistencies between log transprots
