+++ c-sharp/Program.cs
@@ -17,7 +17,11 @@ public static class Program
         Console.WriteLine("Paranext data provider starting up");
         Thread.CurrentThread.Name = "Main";
 
-        var listener = new ConsoleTraceListener { TraceOutputOptions = TraceOptions.DateTime };
+        var listener = new ConsoleTraceListener
+        {
+            TraceOutputOptions = TraceOptions.DateTime,
+            Filter = new TraceExclusionFilter("CreateSharedProject for {0} ({1})"),
+        };
         // Clear the default listeners to stop Debug.Assert from crashing the app
         Trace.Listeners.Clear();
         // Log all trace messages to the console
