+++ src/shared/data/platform.data.ts
@@ -4,8 +4,11 @@
  */
 export const PLATFORM_NAMESPACE = 'platform';
 
-/** Query string passed to the renderer when starting if it should enable noisy dev mode */
-export const DEV_MODE_RENDERER_INDICATOR = '?noisyDevMode';
+/** Query parameter passed to the renderer when starting if it should enable noisy dev mode */
+export const LOG_LEVEL_QUERY_PARAMETER = 'logLevel';
+
+/** Query parameter passed to the renderer when starting if it should enable noisy dev mode */
+export const DEV_MODE_QUERY_PARAMETER = 'noisyDevMode';
 
 /** ID of the default theme family for use in the application */
 export const DEFAULT_THEME_FAMILY = '';
